fix(Record): reorganize edge-case tests
