+++ pkg/gotest/r_test.go
@@ -47,3 +47,32 @@ func TestRecord_ErrorfThenFailNow(t *testing.T) {
 		t.Errorf("message = %q, want %q", rec.Message(), "first error")
 	}
 }
+
+func TestRecord_MultipleErrorf_LastWins(t *testing.T) {
+	rec := Record(func(r *R) {
+		r.Errorf("first")
+		r.Errorf("second")
+		r.Errorf("third")
+	})
+	True(t, rec.Failed())
+	Equal(t, "third", rec.Message())
+}
+
+func TestRecord_FailNowStopsExecution(t *testing.T) {
+	reached := false
+	rec := Record(func(r *R) {
+		r.FailNow()
+		reached = true
+	})
+	True(t, rec.Failed())
+	False(t, reached)
+	Equal(t, "", rec.Message())
+}
+
+func TestRecord_EmptyErrorf(t *testing.T) {
+	rec := Record(func(r *R) {
+		r.Errorf("")
+	})
+	True(t, rec.Failed())
+	Equal(t, "", rec.Message())
+}
